More flexible inter-actor sync primitive
Sync service is merely a central storage node for actor votes.

At each synchronization points, actors can submit one vote.
The range of valid values for this vote is uint16.

At each synchronization points, actors can retrieve the set
of votes submitted to the sync service.

"Validation" of the ballot is done on the client side.
Each client can independently decide to move past the
synchronization point once the set of submitted votes
satisfy its expectations.

A key motivation for this new primitive is the change of
behaviors in the Phoenix codepath that result in asymmetric
states when conflicts are involved:
  - content conflicts are purely local
  - aliased objects are purely local

Change-Id: Id458d4686e887eb01e287b719700c74ab8c16a81